+++ src/Table/index.tsx
@@ -97,7 +97,7 @@ export const ColoredTable = styled(Table)`
   /* Show the header in deep blue */
   th.mll-ant-table-cell {
     background-color: ${(props) => props.theme.tableHeaderBackgroundColor};
-    color: white;
+    color: ${(props) => props.theme.tableHeaderFontColor};
   }
 
   /* Less jarring than the default hover color */
@@ -129,8 +129,16 @@ export const ColoredTable = styled(Table)`
   /* Stripe rows */
 
   /* Depending on the table layout, the header may or may not count as a row */
-  .mll-ant-table.mll-ant-table-fixed-header tr:nth-child(odd) td,
-  .mll-ant-table:not(.mll-ant-table-fixed-header) tr:nth-child(even) td {
+  .mll-ant-table.mll-ant-table-fixed-header
+    tr:not(.mll-ant-table-expanded-row):nth-child(
+      odd of tr:not(.mll-ant-table-expanded-row)
+    )
+    > td,
+  .mll-ant-table:not(.mll-ant-table-fixed-header)
+    tr:not(.mll-ant-table-expanded-row):nth-child(
+      even of tr:not(.mll-ant-table-expanded-row)
+    )
+    > td {
     background-color: ${(props) => props.theme.tableRowStripeBackgroundColor};
   }
 
@@ -187,7 +195,14 @@ export const ColoredTable = styled(Table)`
   }
 
   /* Highlight selected row and avoid visibility of rowSelection checkbox  */
-  .mll-ant-table tr.mll-ant-table-row.mll-ant-table-row-selected td {
+  .mll-ant-table
+    table 
+    tr:not(
+      .mll-ant-table-expanded-row
+    ):nth-child(
+      n /* to be more specific than "odd" or "even" above */
+    ).mll-ant-table-row.mll-ant-table-row-selected
+    > td.mll-ant-table-cell {
     background-color: ${(props) => props.theme.focusedRowColor};
   }
 ` as typeof Table;